# Docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShelfPilot API
  version: 0.8.0
  description: |
    UI + mock API contract for ShelfPilot store layout design (demo stack).
    Auth is mock email/password with role selection. Persistence is SQLite.
    v0.3 adds shelves, separate aisle/shelf mapping, and dimension-based planogram
    (change: layout-editor-planogram).
    v0.4 adds polygon containment, rules-based autogenerate, and category-gated planogram
    (change: layout-autogen-walkthrough).
    v0.5 adds product PATCH, polygon-tight packer, multi-level planogram UX
    (change: merch-layers-polygon-fix).
    v0.6 adds portfolio analytics, smart category-mix autogenerate, temperatureZone on shelves
    (change: module-reframe-smart-autogen).
    v0.7 adds dual-face shelves (faces A/B), displayNumber, strict polygon fixture zone,
    and faceId on planogram writes (change: dual-face-numbered-shelves-strict-polygon).
    v0.8 adds shelf bay segments (horizontal splits), segment-scoped planogram preview/placement,
    and documents the visual planogram editor contract (changes: layout-dimensions-rotation-shelf-bays,
    shelf-planogram-visual-editor).
servers:
  - url: http://localhost:3000
  - url: http://localhost:3001
    description: Docker Compose host-mapped API port
...
paths:
...
  /categories:
    get:
      tags: [Catalog]
      operationId: listCategories
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vertical
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Category tree
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
    post:
      tags: [Catalog]
      operationId: createCategory
      description: Roles — Designer, Admin. `id` optional.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          $ref: "#/components/responses/Forbidden"
  /products:
    get:
      tags: [Catalog]
      operationId: listProducts
      security: [{ bearerAuth: [] }]
      parameters:
        - name: categoryId
          in: query
          schema: { type: string }
        - name: vertical
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
    post:
      tags: [Catalog]
      operationId: createProduct
      description: Roles — Designer, Admin. `id` optional.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          $ref: "#/components/responses/Forbidden"
  /products/{productId}:
    patch:
      tags: [Catalog]
      operationId: updateProduct
      description: Roles — Designer, Admin. Update name, sku, categoryId, attributes.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /catalog/import:
    post:
      tags: [Catalog]
      operationId: importCatalog
      description: Roles — Admin, Designer
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categories:
                  type: array
                  items:
                    $ref: "#/components/schemas/CategoryInput"
                products:
                  type: array
                  items:
                    $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  importedCategories: { type: integer }
                  importedProducts: { type: integer }
        "403":
          $ref: "#/components/responses/Forbidden"
  /catalog/export:
    get:
      tags: [Catalog]
      operationId: exportCatalog
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vertical
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Catalog JSON export
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque session token from POST /auth/login. Expires when AUTH_SESSION_TTL > 0.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    CategoryInput:
      type: object
      required: [name, vertical]
      properties:
        id: { type: string }
        name: { type: string }
        vertical: { type: string }
        parentId: { type: string }
        color: { type: string }
    Category:
      allOf:
        - $ref: "#/components/schemas/CategoryInput"
        - type: object
          required: [id, name, vertical]
    ProductInput:
      type: object
      required: [name, categoryId]
      properties:
        id: { type: string }
        name: { type: string }
        sku: { type: string }
        categoryId: { type: string }
        attributes:
          type: object
          additionalProperties: true
          description: |
            Planogram uses attributes.widthMeters and attributes.heightMeters when present.
            Defaults if missing: width 0.2m, height 0.25m.
    ProductPatch:
      type: object
      properties:
        name: { type: string }
        sku: { type: string }
        categoryId: { type: string }
        attributes:
          type: object
          additionalProperties: true
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
        - type: object
          required: [id, name, categoryId]
...